+++ cardinal-components-level/src/main/java/org/ladysnake/cca/internal/level/CardinalComponentsLevel.java
@@ -23,41 +23,60 @@
 package org.ladysnake.cca.internal.level;
 
 import com.mojang.datafixers.util.Unit;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLevelEvents;
 import net.fabricmc.loader.api.FabricLoader;
-import net.minecraft.network.protocol.common.ClientboundCustomPayloadPacket;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.world.level.storage.LevelData;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.Nullable;
+import org.ladysnake.cca.api.v3.component.ComponentContainer;
 import org.ladysnake.cca.api.v3.component.ComponentKey;
 import org.ladysnake.cca.api.v3.component.ComponentProvider;
-import org.ladysnake.cca.api.v3.component.sync.AutoSyncedComponent;
-import org.ladysnake.cca.api.v3.world.WorldSyncCallback;
+import org.ladysnake.cca.api.v8.level.LevelSyncCallback;
 import org.ladysnake.cca.internal.base.ComponentUpdatePayload;
 import org.ladysnake.cca.internal.base.MorePacketCodecs;
 
-public final class CardinalComponentsLevel {
+import java.util.IdentityHashMap;
+import java.util.Map;
 
-    /**
-     * {@link ClientboundCustomPayloadPacket} channel for default level component synchronization.
-     *
-     * <p> Components synchronized through this channel will have {@linkplain AutoSyncedComponent#applySyncPacket(net.minecraft.network.RegistryFriendlyByteBuf)}
-     * called on the game thread.
-     */
-    public static final CustomPacketPayload.Type<ComponentUpdatePayload<Unit>> PACKET_ID = ComponentUpdatePayload.id("level_sync");
+public final class CardinalComponentsLevel {
+    public static final CustomPacketPayload.Type<ComponentUpdatePayload<Unit>> PACKET_ID = ComponentUpdatePayload.id("world_sync");
+    private static final Map<@Nullable ResourceKey<Level>, ComponentContainer.Factory<Level>> levelContainerFactories = new IdentityHashMap<>();
 
     public static void init() {
         if (FabricLoader.getInstance().isModLoaded("fabric-networking-api-v1")) {
             ComponentUpdatePayload.register(PACKET_ID, MorePacketCodecs.EMPTY);
-            if (FabricLoader.getInstance().isModLoaded("cardinal-components-world")) {
-                WorldSyncCallback.EVENT.register((player, world) -> {
-                    LevelData props = world.getLevelData();
-
-                    for (ComponentKey<?> key : ((ComponentProvider) props).getComponentContainer().keys()) {
-                        key.syncWith(player, ((ComponentProvider) props));
-                    }
-                });
-            }
+            LevelSyncCallback.EVENT.register((player, world) -> {
+                ComponentProvider provider = (ComponentProvider) world;
+                for (ComponentKey<?> key : provider.getComponentContainer().keys()) {
+                    key.syncWith(player, provider);
+                }
+            });
+        }
+        if (FabricLoader.getInstance().isModLoaded("fabric-lifecycle-events-v1")) {
+            ServerLevelEvents.LOAD.register((server, world) -> ((ComponentProvider) world).getComponentContainer().onServerLoad());
+            ServerLevelEvents.UNLOAD.register((server, world) -> ((ComponentProvider) world).getComponentContainer().onServerUnload());
         }
         StaticLevelComponentPlugin.INSTANCE.ensureInitialized();
     }
 
+    public static ComponentContainer createComponents(Level world) {
+        return levelContainerFactories.computeIfAbsent(
+            world.dimension(),
+            CardinalComponentsLevel::getLevelComponentFactory
+        ).createContainer(world);
+    }
+
+
+    private static synchronized ComponentContainer.Factory<Level> getLevelComponentFactory(ResourceKey<Level> dimensionKey) {
+        // need to check again despite synchronization, because the factory may have been generated while waiting from createComponents
+        ComponentContainer.Factory<Level> existing = levelContainerFactories.get(dimensionKey);
+        if (existing != null) return existing;
+
+        if (StaticLevelComponentPlugin.INSTANCE.requiresStaticFactory(dimensionKey)) {
+            return StaticLevelComponentPlugin.INSTANCE.buildDedicatedFactory(dimensionKey);
+        } else {
+            return levelContainerFactories.computeIfAbsent(null, n -> StaticLevelComponentPlugin.INSTANCE.buildDedicatedFactory(null));
+        }
+    }
 }