+++ cardinal-components-leveldata/src/testmod/java/org/ladysnake/cca/test/leveldata/CcaLevelDataTestSuite.java
@@ -1,0 +1,41 @@
+/*
+ * Cardinal-Components-API
+ * Copyright (C) 2019-2026 Ladysnake
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+ * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
+ * OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ladysnake.cca.test.leveldata;
+
+import net.fabricmc.fabric.api.gametest.v1.GameTest;
+import net.minecraft.gametest.framework.GameTestHelper;
+import net.minecraft.network.chat.Component;
+import org.ladysnake.cca.api.v3.component.ComponentAccess;
+import org.ladysnake.cca.test.base.TickingTestComponent;
+
+public class CcaLevelDataTestSuite {
+    @GameTest
+    public void levelComponentsTick(GameTestHelper ctx) {
+        int baseTicks = ((ComponentAccess) ctx.getLevel().getLevelData()).getComponent(TickingTestComponent.KEY).serverTicks();
+        ctx.runAfterDelay(5, () -> {
+            int ticks = ((ComponentAccess) ctx.getLevel().getLevelData()).getComponent(TickingTestComponent.KEY).serverTicks();
+            ctx.assertValueEqual(5, ticks - baseTicks, Component.literal("Component should tick 5 times -"));
+            ctx.succeed();
+        });
+    }
+}