+++ changelog.md
@@ -1,10 +1,18 @@
 ------------------------------------------------------
-Version 8.0.0-alpha.3+26.1-pre-2
+Version 8.0.0-alpha.3+26.1-pre-3
 ------------------------------------------------------
-Updated to 26.1-pre-2
+Updated to 26.1
+
+See the migration guide available on [the Ladysnake website](https://ladysnake.org/wiki/cardinal-components-api/upgrade-instructions/CCA-8-changes).
+
+**Changes**
+- Two modules have been renamed in order to match Mojmap names :
+  - `cardinal-components-level` has been renamed to `cardinal-components-leveldata`
+  - `cardinal-components-world` has been renamed to `cardinal-components-level`
+  - All interfaces inside these modules have been renamed accordingly
 
 **Removed**
-- Components attached to a `LevelData` object (using the `cardinal-components-level` module) can no longer use dynamic registries in serialization (vanilla limitation)
+- Components attached to a `LevelData` object (using the `cardinal-components-leveldata` module) can no longer use dynamic registries in serialization (vanilla limitation)
 
 ------------------------------------------------------
 Version 7.3.1